requirements.txt
pdfplumber==0.10.2
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
nltk==3.8.1
spacy==3.7.4
streamlit==1.35.0